thermal: Add simple thermal.
Based on custom thermal driver by Francisco Franco.

This enables thermal monitoring capability in the kernel in the absence of a system wide thermal monitoring entity. Monitors TSENS temperature and limits the max frequency of the cores. The code is simple, uses 3 throttling steps, and is completely configurable via sysfs.

Signed-off-by: AyushR1 <ayushrathore12501@gmail.com>